Add EntityMetaData.hasAttributeExpression(attrName), merge duplicate test classes
