+++ molgenis-data/src/main/java/org/molgenis/data/support/DefaultEntityMetaData.java
@@ -1,5 +1,7 @@
 package org.molgenis.data.support;
 
+import static java.util.stream.StreamSupport.stream;
+
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -130,8 +132,8 @@ public class DefaultEntityMetaData extends AbstractEntityMetaData implements Edi
 	public void addAttributeMetaData(AttributeMetaData attributeMetaData)
 	{
 		if (attributeMetaData == null) throw new IllegalArgumentException("AttributeMetaData cannot be null");
-		if (attributeMetaData.getName() == null) throw new IllegalArgumentException(
-				"Name of the AttributeMetaData cannot be null");
+		if (attributeMetaData.getName() == null)
+			throw new IllegalArgumentException("Name of the AttributeMetaData cannot be null");
 		if (attributeMetaData.isLabelAttribute()) setLabelAttribute(attributeMetaData.getName());
 		if (attributeMetaData.isIdAtrribute()) setIdAttribute(attributeMetaData.getName());
 
@@ -236,8 +238,8 @@ public class DefaultEntityMetaData extends AbstractEntityMetaData implements Edi
 		strBuilder.append(this.getName()).append('\'');
 		if (isAbstract()) strBuilder.append(" abstract='true'");
 		if (getExtends() != null) strBuilder.append(" extends='" + getExtends().getName()).append('\'');
-		if (getIdAttribute() != null) strBuilder.append(" idAttribute='").append(getIdAttribute().getName())
-				.append('\'');
+		if (getIdAttribute() != null)
+			strBuilder.append(" idAttribute='").append(getIdAttribute().getName()).append('\'');
 		if (getDescription() != null) strBuilder.append(" description='")
 				.append(getDescription().substring(0, Math.min(25, getDescription().length())))
 				.append(getDescription().length() > 25 ? "...'" : "'");
@@ -271,6 +273,12 @@ public class DefaultEntityMetaData extends AbstractEntityMetaData implements Edi
 		}
 		else if (!getName().equals(other.getName())) return false;
 		return true;
+	}
+
+	@Override
+	public boolean hasAttributeWithExpression()
+	{
+		return stream(getAtomicAttributes().spliterator(), false).anyMatch(attr -> attr.getExpression() != null);
 	}
 
 }
